% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{basemap_kor_in}
\alias{basemap_kor_in}
\title{Basemap layout for \code{tmap} in South Korea}
\usage{
basemap_kor_in(year = 2020)
}
\arguments{
\item{year}{An integer specifying
the year of the administrative level 2 data.
Default is 2020.}
}
\value{
A \code{tmap} object representing the basemap layout for South Korea.
}
\description{
This function first downloads adjacent country polygon data from GADM
and then creates a basemap layout with insets for South Korea.
The sf objects are reprojected to EPSG:5179 (Korea 2000 / Unified CS).
}
\note{
The output is a tmap object that can be further customized.
It is recommended to export the map using \code{tmap::tmap_save()}
with a high resolution (i.e., dpi > 200).
}
