% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_basic_adm2.R
\name{convert_basic_adm2}
\alias{convert_basic_adm2}
\title{Convert administrative level 2 sf polygons to basic local government boundaries}
\usage{
convert_basic_adm2(adm2_sf = NULL, year = 2020)
}
\arguments{
\item{adm2_sf}{An \code{sf} object containing administrative level 2 polygons from Ministry of Data and Statistics.}

\item{year}{An integer specifying the year of the administrative level 2 data. Default is 2020. This value is ignored when \code{adm2_sf} is provided.}
}
\value{
An \code{sf} object containing basic local government boundaries.
}
\description{
This function converts administrative level 2 polygons from Ministry of Data and Statistics (formerly Statistics Korea) to basic local government boundaries. Basic local governments are administrative units with their own local councils and mayors, such as cities (si), counties (gun) and districts (gu), but without districts in populous non-metropolitan cities.
}
