% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_districts.R
\name{load_districts}
\alias{load_districts}
\title{Load \code{sf} district polygons}
\usage{
load_districts(year = 2020)
}
\arguments{
\item{year}{integer(1). census year to load. Defaults to 2020.}
}
\value{
An \code{sf} object containing the district polygons.
}
\description{
Load \code{sf} district polygons
}
\details{
The function loads district polygons for the specified year from the package's
extdata directory. The polygons are stored in an RDS file and are read using
the \code{readRDS} function. The polygons are in the Simple Features (sf) format,
which is suitable for spatial data analysis in R. The polygons include following
attributes:
\itemize{
\item \code{year}: The census year (e.g., 2020).
\item \code{adm2_code}: The administrative code for the district.
}
}
\examples{
library(sf)
sf_use_s2(FALSE)
sf_2020 <- load_districts(year = 2020)
}
